requirements.txt
python-telegram-bot
sc2reader
pdfkit
gitpython
zephyrus_sc2_parser
spawningtool
numpy
